+++ src/did.js
@@ -1,0 +1,112 @@
+import { IonDid, IonRequest, LocalSigner } from "@decentralized-identity/ion-sdk";
+import { generateKeyPair } from "./utils";
+
+export class DID {
+  constructor(options = {}) {
+    this._ops = options.ops || [];
+    if (!this._ops[0]) {
+      this._ops[0] = this.generateOperation('create', options.content || {}, false);
+    }
+  }
+
+  async generateOperation(type, content, commit = true) {
+    let ops = await this.getAllOperations();
+    let lastOp = ops[ops.length - 1];
+    if (lastOp && lastOp.operation === 'deactivate') {
+      throw 'Cannot perform further operations on a deactivated DID'
+    }
+    let op = {
+      operation: type,
+      content: content
+    };
+    if (type !== 'create') {
+      op.previous = ops.reduce((last, op) => {
+        return op.operation === type || (op.operation === 'recover' && (type === 'deactivate' || type === 'update')) ? op : last;
+      }, ops[0])
+    }
+    if (type === 'create' || type === 'recover') {
+      op.recovery = await generateKeyPair();
+    }
+    if (type !== 'deactivate') {
+      op.update = await generateKeyPair();
+    }
+    if (commit) this._ops.push(op);
+    return op;
+  }
+
+  async generateRequest(payload = 0, options = {}) {
+    const op = typeof payload === 'number' ? await this.getOperation(payload) : payload;
+
+    switch (op.operation) {
+      case 'update':
+        return IonRequest.createUpdateRequest({
+          didSuffix: await this.getSuffix(),
+          signer: options.signer || LocalSigner.create(op.previous.update.privateJwk),
+          updatePublicKey: op.previous.update.publicJwk,
+          nextUpdatePublicKey: op.update.publicJwk,
+          servicesToAdd: op.content?.addServices,
+          idsOfServicesToRemove: op.content?.removeServices,
+          publicKeysToAdd: op.content?.addPublicKeys,
+          idsOfPublicKeysToRemove: op.content?.removePublicKeys
+        });
+        break;
+
+      case 'recover':
+        return IonRequest.createRecoverRequest({
+          didSuffix: await this.getSuffix(),
+          signer: options.signer || LocalSigner.create(op.previous.recovery.privateJwk),
+          recoveryPublicKey: op.previous.recovery.publicJwk,
+          nextRecoveryPublicKey: op.recovery.publicJwk,
+          nextUpdatePublicKey: op.update.publicJwk,
+          document: op.content
+        });
+
+      case 'deactivate':
+        return IonRequest.createDeactivateRequest({
+          didSuffix: await this.getSuffix(),
+          recoveryPublicKey: op.previous.recovery.publicJwk,
+          signer: options.signer || LocalSigner.create(op.previous.recovery.privateJwk)
+        });
+
+      case 'create':
+      default:
+        return IonRequest.createCreateRequest({
+          recoveryKey: op.recovery.publicJwk,
+          updateKey: op.update.publicJwk,
+          document: op.content
+        });
+    }
+  }
+
+  getAllOperations() {
+    return Promise.all(this._ops);
+  }
+
+  async getOperation(index) {
+    return this._ops[index];
+  }
+
+  async getState() {
+    return {
+      shortForm: await this.getURI('short'),
+      longForm: await this.getURI(),
+      ops: await this.getAllOperations()
+    }
+  }
+
+  async getSuffix() {
+    const uri = await this.getURI('short');
+    return uri.split(':').pop();
+  }
+
+  async getURI(form) {
+    const create = await this.getOperation(0);
+    this._longForm = this._longForm || await IonDid.createLongFormDid({
+      recoveryKey: create.recovery.publicJwk,
+      updateKey: create.update.publicJwk,
+      document: create.content
+    });
+
+    return !form || form === 'long' ? this._longForm : this._longForm.split(':').slice(0, -1).join(':');
+  }
+}
