+++ src/index.js
@@ -1,0 +1,2 @@
+export { DID } from './did';
+export { anchor, generateKeyPair, sign, verify } from './utils';
